norminette

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: eassouli <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 19:17:52 by eassouli          #+#    #+#             */
-/*   Updated: 2019/10/27 14:49:02 by eassouli         ###   ########.fr       */
+/*   Updated: 2019/10/27 15:11:02 by eassouli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int		max;
+	size_t	max;
 	size_t	i;
 	size_t	dstlen;
 	size_t	srclen;
